// next-cms/api/dto/lending.dto.ts
export interface LendingDTO {
	id: number;

	headerName: string;

	headerDescription: string;

	mainImg: string;

	navIntroduction: string;

	introductionTitle: string;

	introductionDescription: string;

	navAbout: string;

	aboutTitle: string;

	aboutDescription: string;

	navContact: string;

	contactTitle: string;

	contactDescription: string;

	footerCompany: string;

	useHeader: boolean;

	useMainImg: boolean;

	useNavIntroduction: boolean;

	useNavAbout: boolean;

	useNavContact: boolean;

	useFooter: boolean;

	userId: number;
}